Removed selenium and custom request crawler implementation. Made scrapy as default cralwer
