+++ backend/src/crawlers/crawler_factory.py
@@ -12,21 +12,7 @@ from .interface import (
     CrawlerManager
 )
 
-from .impl import (
-    CrawlerFactory,
-    create_and_configure_scrapy_crawler,
-    create_and_configure_selenium_crawler,
-    test_crawler_creation
-)
-
-from src.crawlers.implementations import (
-    ScrapyCrawler,
-    SeleniumCrawler,
-    SCRAPY_CONFIG_EXAMPLE,
-    SELENIUM_CONFIG_EXAMPLE,
-    create_scrapy_crawler,
-    create_selenium_crawler
-)
+from .scrapy_crawler import ScrapyCrawler, create_scrapy_crawler
 
 import logging
 from typing import Dict, Any, Optional, List
@@ -44,7 +30,6 @@ def get_crawler_manager() -> CrawlerManager:
 
 
 def create_and_register_crawler(
-    crawler_type: str, 
     name: str, 
     config: Dict[str, Any],
     auto_register: bool = True
@@ -53,7 +38,6 @@ def create_and_register_crawler(
     Create a crawler and optionally register it with the global manager
     
     Args:
-        crawler_type: Type of crawler ('scrapy', 'selenium', etc.)
         name: Unique name for the crawler
         config: Configuration dictionary
         auto_register: Whether to automatically register with global manager
@@ -61,111 +45,7 @@ def create_and_register_crawler(
     Returns:
         CrawlerInterface: Created crawler instance
     """
-    crawler = CrawlerFactory.create_crawler(crawler_type, name, config)
-    
-    if auto_register:
-        crawler_manager.add_crawler(crawler)
-    
-    return crawler
-
-
-def get_supported_crawler_types() -> List[str]:
-    """Get list of all supported crawler types"""
-    return CrawlerFactory.get_supported_types()
-
-
-def register_custom_crawler_type(crawler_type: str, crawler_class):
-    """
-    Register a custom crawler type
-    
-    Args:
-        crawler_type: String identifier for the crawler type
-        crawler_class: Class implementing CrawlerInterface
-    """
-    CrawlerFactory.register_crawler_type(crawler_type, crawler_class)
-
-
-def get_crawler_type_info(crawler_type: str) -> Dict[str, Any]:
-    """
-    Get information about a specific crawler type
-    
-    Args:
-        crawler_type: Type of crawler
-        
-    Returns:
-        Dict containing crawler information
-    """
-    return CrawlerFactory.get_crawler_info(crawler_type)
-
-
-def list_all_crawler_types_info() -> Dict[str, Dict[str, Any]]:
-    """Get detailed information about all registered crawler types"""
-    return CrawlerFactory.list_all_crawler_info()
-
-
-# Convenience functions for creating specific crawler types
-def create_scrapy_crawler_with_settings(
-    name: str,
-    spider_name: str,
-    custom_settings: Dict[str, Any] = None,
-    auto_register: bool = True,
-    **kwargs
-) -> ScrapyCrawler:
-    """
-    Create a Scrapy crawler with custom settings and optionally register it
-    
-    Args:
-        name: Crawler name
-        spider_name: Name of the Scrapy spider
-        custom_settings: Custom Scrapy settings to override defaults
-        auto_register: Whether to register with global manager
-        **kwargs: Additional configuration options
-        
-    Returns:
-        ScrapyCrawler: Configured Scrapy crawler instance
-    """
-    crawler = create_and_configure_scrapy_crawler(
-        name=name,
-        spider_name=spider_name,
-        settings=custom_settings,
-        **kwargs
-    )
-    
-    if auto_register:
-        crawler_manager.add_crawler(crawler)
-    
-    return crawler
-
-
-def create_selenium_crawler_with_options(
-    name: str,
-    driver_type: str = 'chrome',
-    headless: bool = True,
-    custom_options: Dict[str, Any] = None,
-    auto_register: bool = True,
-    **kwargs
-) -> SeleniumCrawler:
-    """
-    Create a Selenium crawler with custom options and optionally register it
-    
-    Args:
-        name: Crawler name
-        driver_type: Type of WebDriver ('chrome', 'firefox', etc.)
-        headless: Whether to run in headless mode
-        custom_options: Custom driver options
-        auto_register: Whether to register with global manager
-        **kwargs: Additional configuration options
-        
-    Returns:
-        SeleniumCrawler: Configured Selenium crawler instance
-    """
-    crawler = create_and_configure_selenium_crawler(
-        name=name,
-        driver_type=driver_type,
-        headless=headless,
-        custom_options=custom_options,
-        **kwargs
-    )
+    crawler = create_scrapy_crawler(name, **config)
     
     if auto_register:
         crawler_manager.add_crawler(crawler)
@@ -316,69 +196,6 @@ def get_running_crawler_names() -> List[str]:
     return list(crawler_manager.get_running_crawlers().keys())
 
 
-# Example usage and demo functions
-def demo_crawler_usage():
-    """Demonstrate crawler usage with examples"""
-    print("=== Crawler Factory Demo ===")
-    
-    try:
-        # Create Scrapy crawler
-        scrapy_crawler = create_scrapy_crawler_with_settings(
-            name="demo_scrapy",
-            spider_name="quotes_spider",
-            custom_settings={
-                'DOWNLOAD_DELAY': 2,
-                'CONCURRENT_REQUESTS': 8
-            }
-        )
-        print(f"Created Scrapy crawler: {scrapy_crawler.name}")
-        
-        # Create Selenium crawler
-        selenium_crawler = create_selenium_crawler_with_options(
-            name="demo_selenium",
-            driver_type="chrome",
-            headless=True,
-            custom_options={
-                'disable_images': True
-            }
-        )
-        print(f"Created Selenium crawler: {selenium_crawler.name}")
-        
-        # Show all registered crawlers
-        print(f"Registered crawlers: {crawler_manager.list_crawlers()}")
-        
-        # Show supported types
-        print(f"Supported crawler types: {get_supported_crawler_types()}")
-        
-        # Show detailed type information
-        print("Crawler type information:")
-        for crawler_type, info in list_all_crawler_types_info().items():
-            print(f"  {crawler_type}: {info['class_name']}")
-        
-        # Test starting and stopping
-        print("\n=== Testing Crawler Operations ===")
-        
-        # Start Selenium crawler (Scrapy requires actual spider setup)
-        if start_crawler_by_name("demo_selenium"):
-            print("Started demo_selenium crawler")
-            
-            # Get status
-            status = get_crawler_status_by_name("demo_selenium")
-            if status:
-                print(f"Selenium crawler status: {status['status']}")
-            
-            # Stop crawler
-            if stop_crawler_by_name("demo_selenium"):
-                print("Stopped demo_selenium crawler")
-        
-        print("Demo completed successfully!")
-        return True
-        
-    except Exception as e:
-        logger.error(f"Demo failed: {str(e)}")
-        return False
-
-
 # Export main components
 __all__ = [
     # Interfaces and enums
@@ -388,16 +205,11 @@ __all__ = [
     'CrawlerManager',
     
     # Factory and implementations
-    'CrawlerFactory',
     'ScrapyCrawler',
-    'SeleniumCrawler',
     
     # Utility functions
     'create_scrapy_crawler',
-    'create_selenium_crawler',
     'create_and_register_crawler',
-    'create_scrapy_crawler_with_settings',
-    'create_selenium_crawler_with_options',
     
     # Manager functions
     'get_crawler_manager',
@@ -413,25 +225,6 @@ __all__ = [
     'get_all_crawler_statuses',
     'get_running_crawler_names',
     
-    # Information functions
-    'get_supported_crawler_types',
-    'get_crawler_type_info',
-    'list_all_crawler_types_info',
-    'register_custom_crawler_type',
-    
-    # Configuration examples
-    'SCRAPY_CONFIG_EXAMPLE',
-    'SELENIUM_CONFIG_EXAMPLE',
-    
     # Global manager
     'crawler_manager',
-    
-    # Demo and testing
-    'demo_crawler_usage',
-    'test_crawler_creation'
 ]
-
-
-# if __name__ == "__main__":
-    # Run demo
-    # asyncio.run(demo_crawler_usage())
